chore: regression gen results and validation (#183)
