+++ .env.example
@@ -3,3 +3,5 @@ DATABASE_PORT=5432
 DATABASE_NAME=coredb
 DATABASE_USER=postgres
 DATABASE_PASSWORD=postgres
+REGRESSION_PATH="/path/to/your/CDISC_Sharepoint_dump/unitTesting/"
+CURRENT_RULE_DEV="CORE-XXXXXX"
