+++ .prettierignore
@@ -1,0 +1,1 @@
+tests/resources/rules/rules.json
