+++ cdisc_rules_engine/models/validation_error_entity.py
@@ -31,12 +31,15 @@ class ValidationErrorEntity(BaseValidationEntity):
         data = {}
         for key, val in self.value.items():
             if isinstance(val, set):
-                data[key] = list(val)
+                data[key] = sorted(list(val))
+            elif isinstance(val, list):
+                data[key] = sorted(val)
             elif isinstance(val, BaseEnum):
                 data[key] = val.value
             else:
                 data[key] = val
-        return data
+        sorted_data = {k: data[k] for k in sorted(data.keys())}
+        return sorted_data
 
     def to_representation(self) -> dict:
         representation: dict = {
