+++ cdisc_rules_engine/rules_engine.py
@@ -78,9 +78,7 @@ class RulesEngine:
         )
         self.dataset_implementation = data_service_factory.get_dataset_implementation()
         kwargs["dataset_implementation"] = self.dataset_implementation
-        self.data_service = data_service or data_service_factory.get_data_service(
-            self.dataset_paths
-        )
+        self.data_service = data_service or data_service_factory.get_data_service(self.dataset_paths)
         self.rule_processor = RuleProcessor(
             self.data_service,
             self.cache,
@@ -98,9 +96,7 @@ class RulesEngine:
 
     def validate_single_rule(self, rule: dict, datasets: Iterable[SDTMDatasetMetadata]):
         results = {}
-        rule["conditions"] = ConditionCompositeFactory.get_condition_composite(
-            rule["conditions"]
-        )
+        rule["conditions"] = ConditionCompositeFactory.get_condition_composite(rule["conditions"])
         for dataset_metadata in datasets:
             if dataset_metadata.unsplit_name in results and "domains" in rule:
                 include_split = rule["domains"].get("include_split_datasets", False)
@@ -136,12 +132,8 @@ class RulesEngine:
                 self.standard_substandard,
             )
             if is_suitable:
-                result: List[Union[dict, str]] = self.validate_rule(
-                    rule, datasets, dataset_metadata
-                )
-                logger.info(
-                    f"Validated dataset {dataset_metadata.name}. Result = {result}"
-                )
+                result: List[Union[dict, str]] = self.validate_rule(rule, datasets, dataset_metadata)
+                logger.info(f"Validated dataset {dataset_metadata.name}. Result = {result}")
                 if result:
                     return result
                 else:
@@ -150,16 +142,13 @@ class RulesEngine:
                         ValidationErrorContainer(
                             **{
                                 "dataset": dataset_metadata.filename,
-                                "domain": dataset_metadata.domain
-                                or dataset_metadata.rdomain,
+                                "domain": dataset_metadata.domain or dataset_metadata.rdomain,
                                 "errors": [],
                             }
                         ).to_representation()
                     ]
             else:
-                logger.info(
-                    f"Skipped dataset {dataset_metadata.name}. Reason: {reason}"
-                )
+                logger.info(f"Skipped dataset {dataset_metadata.name}. Reason: {reason}")
                 error_obj: ValidationErrorContainer = ValidationErrorContainer(
                     status=ExecutionStatus.SKIPPED.value,
                     message=reason,
@@ -226,61 +215,35 @@ class RulesEngine:
         # SPECIAL CASES FOR RULE TYPES ###############################
         # TODO: Handle these special cases better.
         if self.library_metadata:
-            kwargs["variable_codelist_map"] = (
-                self.library_metadata.variable_codelist_map
-            )
-            kwargs["codelist_term_maps"] = (
-                self.library_metadata.get_all_ct_package_metadata()
-            )
+            kwargs["variable_codelist_map"] = self.library_metadata.variable_codelist_map
+            kwargs["codelist_term_maps"] = self.library_metadata.get_all_ct_package_metadata()
         if rule.get("rule_type") == RuleTypes.DEFINE_ITEM_METADATA_CHECK.value:
             if self.library_metadata:
-                kwargs["variable_codelist_map"] = (
-                    self.library_metadata.variable_codelist_map
-                )
-                kwargs["codelist_term_maps"] = (
-                    self.library_metadata.get_all_ct_package_metadata()
-                )
+                kwargs["variable_codelist_map"] = self.library_metadata.variable_codelist_map
+                kwargs["codelist_term_maps"] = self.library_metadata.get_all_ct_package_metadata()
         elif (
-            rule.get("rule_type")
-            == RuleTypes.VARIABLE_METADATA_CHECK_AGAINST_DEFINE.value
-            or rule.get("rule_type")
-            == RuleTypes.VARIABLE_METADATA_CHECK_AGAINST_DEFINE_XML_AND_LIBRARY.value
+            rule.get("rule_type") == RuleTypes.VARIABLE_METADATA_CHECK_AGAINST_DEFINE.value
+            or rule.get("rule_type") == RuleTypes.VARIABLE_METADATA_CHECK_AGAINST_DEFINE_XML_AND_LIBRARY.value
         ):
-            self.rule_processor.add_comparator_to_rule_conditions(
-                rule, comparator=None, target_prefix="define_"
-            )
-        elif (
-            rule.get("rule_type")
-            == RuleTypes.VALUE_LEVEL_METADATA_CHECK_AGAINST_DEFINE.value
-        ):
+            self.rule_processor.add_comparator_to_rule_conditions(rule, comparator=None, target_prefix="define_")
+        elif rule.get("rule_type") == RuleTypes.VALUE_LEVEL_METADATA_CHECK_AGAINST_DEFINE.value:
             value_level_metadata: List[dict] = self.get_define_xml_value_level_metadata(
                 dataset_metadata.full_path, dataset_metadata.unsplit_name
             )
             kwargs["value_level_metadata"] = value_level_metadata
 
-        elif (
-            rule.get("rule_type")
-            == RuleTypes.DATASET_CONTENTS_CHECK_AGAINST_DEFINE_AND_LIBRARY.value
-        ):
-            library_metadata: dict = self.library_metadata.variables_metadata.get(
-                dataset_metadata.domain, {}
-            )
+        elif rule.get("rule_type") == RuleTypes.DATASET_CONTENTS_CHECK_AGAINST_DEFINE_AND_LIBRARY.value:
+            library_metadata: dict = self.library_metadata.variables_metadata.get(dataset_metadata.domain, {})
             define_metadata: List[dict] = builder.get_define_xml_variables_metadata()
-            targets: List[str] = (
-                self.data_processor.filter_dataset_columns_by_metadata_and_rule(
-                    dataset.columns.tolist(), define_metadata, library_metadata, rule
-                )
+            targets: List[str] = self.data_processor.filter_dataset_columns_by_metadata_and_rule(
+                dataset.columns.tolist(), define_metadata, library_metadata, rule
             )
             rule_copy = deepcopy(rule)
-            updated_conditions = RuleProcessor.duplicate_conditions_for_all_targets(
-                rule_copy["conditions"], targets
-            )
+            updated_conditions = RuleProcessor.duplicate_conditions_for_all_targets(rule_copy["conditions"], targets)
             rule_copy["conditions"].set_conditions(updated_conditions)
             # When duplicating conditions,
             # rule should be copied to prevent updates to concurrent rule executions
-            return self.execute_rule(
-                rule_copy, dataset, datasets, dataset_metadata, **kwargs
-            )
+            return self.execute_rule(rule_copy, dataset, datasets, dataset_metadata, **kwargs)
 
         kwargs["ct_packages"] = list(self.ct_packages)
 
@@ -317,9 +280,7 @@ class RulesEngine:
         # Adding copy for now to avoid updating cached dataset
         dataset = deepcopy(dataset)
         # preprocess dataset
-        dataset_preprocessor = DatasetPreprocessor(
-            dataset, dataset_metadata, self.data_service, self.cache
-        )
+        dataset_preprocessor = DatasetPreprocessor(dataset, dataset_metadata, self.data_service, self.cache)
         dataset = dataset_preprocessor.preprocess(rule_copy, datasets)
         dataset = self.rule_processor.perform_rule_operations(
             rule_copy,
@@ -354,9 +315,7 @@ class RulesEngine:
         )
         return results
 
-    def get_define_xml_value_level_metadata(
-        self, dataset_path: str, domain_name: str
-    ) -> List[dict]:
+    def get_define_xml_value_level_metadata(self, dataset_path: str, domain_name: str) -> List[dict]:
         """
         Gets Define XML variable metadata and returns it as dataframe.
         """
@@ -365,9 +324,7 @@ class RulesEngine:
         )
         return define_xml_reader.extract_value_level_metadata(domain_name=domain_name)
 
-    def handle_validation_exceptions(  # noqa
-        self, exception, dataset_path, file_name
-    ) -> ValidationErrorContainer:
+    def handle_validation_exceptions(self, exception, dataset_path, file_name) -> ValidationErrorContainer:  # noqa
         if isinstance(exception, DatasetNotFoundError):
             error_obj = FailedValidationEntity(
                 dataset=os.path.basename(dataset_path),
@@ -390,32 +347,36 @@ class RulesEngine:
             )
             message = "rule execution error"
         elif isinstance(exception, KeyError):
+            message = ", ".join(sorted(exception.args[0].split(", ")))
             error_obj = FailedValidationEntity(
                 dataset=os.path.basename(dataset_path),
                 error="Column not found in data",
-                message=exception.args[0],
+                message=message,
             )
             message = "rule execution error"
         elif isinstance(exception, DomainNotFoundInDefineXMLError):
+            message = ", ".join(sorted(exception.args[0].split(", ")))
             error_obj = FailedValidationEntity(
                 dataset=os.path.basename(dataset_path),
                 error=DomainNotFoundInDefineXMLError.description,
-                message=exception.args[0],
+                message=message,
             )
             message = "rule execution error"
         elif isinstance(exception, VariableMetadataNotFoundError):
+            message = ", ".join(sorted(exception.args[0].split(", ")))
             error_obj = FailedValidationEntity(
                 dataset=os.path.basename(dataset_path),
                 error=VariableMetadataNotFoundError.description,
-                message=exception.args[0],
+                message=message,
             )
             message = "rule execution error"
         elif isinstance(exception, FailedSchemaValidation):
+            message = ", ".join(sorted(exception.args[0].split(", ")))
             if self.validate_xml:
                 error_obj: ValidationErrorContainer = ValidationErrorContainer(
                     status=ExecutionStatus.SKIPPED.value,
                     error=FailedSchemaValidation.description,
-                    message=exception.args[0],
+                    message=message,
                 )
                 message = "Schema Validation Error"
                 errors = [error_obj]
