+++ requirements-dev.txt
@@ -1,0 +1,1 @@
+deepdiff==8.6.0
